+++ branches/4.3.10/xymond/xymond.c
@@ -5006,7 +5006,7 @@ int main(int argc, char *argv[])
 	{
 		FILE *fd = fopen(pidfile, "w");
 		if (fd) {
-			if (fprintf(fd, "%d\n", (int)getpid()) <= 0) {
+			if (fprintf(fd, "%lu\n", (unsigned long)getpid()) <= 0) {
 				errprintf("Error writing PID file %s: %s\n", pidfile, strerror(errno));
 			}
 			fclose(fd);
